refactor: remove EduCardWithTxQueue and keep the Educard and TxQueue separately
