--- apps/bridge-dapp/src/containers/EduCardWithTxQueue/index.ts
@@ -1,1 +1,0 @@
-export { default as EduCardWithTxQueue } from './EduCardWithTxQueue';
