--- apps/bridge-dapp/src/containers/EduCardWithTxQueue/types.ts
@@ -1,12 +1,0 @@
-import { TransactionPayload } from '@webb-tools/webb-ui-components';
-import { BridgeTabType } from '../../types';
-
-export interface EduCardWithTxQueueProps {
-  activeTab: BridgeTabType;
-}
-
-export interface TxQueueContainerProps {
-  isDisplay: boolean;
-
-  transactionPayloads: TransactionPayload[];
-}
